web/requirements.txt
gunicorn==19.4.5
marshmallow==2.1.0
psycopg2==2.6.1
webargs==0.18
Flask==0.10.1
Flask-Cors==2.1.0
Flask-SQLAlchemy==2.0
psycopg2==2.6.1
Werkzeug==0.11.3